Split out gso header
