+++ rust/libs/bin-shared/src/tun_device_manager/linux.rs
@@ -1,5 +1,6 @@
 //! Linux TUN device manager and implementation
 
+mod gso_header;
 mod manager;
 mod tun;
 mod tun_gso_queue;
